refactor: Renamings
- AcceptanceSpec is now the base for all tests (was BaseSpec)
- rename GraphServices -> ApplicationServices
- pulled out DatasetsEncoders
