+++ acceptance-tests/src/test/scala/io/renku/graph/acceptancetests/ReProvisioningSpec.scala
@@ -26,7 +26,7 @@ import io.renku.generators.Generators.Implicits._
 import io.renku.generators.Generators.nonEmptyStrings
 import io.renku.graph.acceptancetests.data._
 import io.renku.graph.acceptancetests.flows.TSProvisioning
-import io.renku.graph.acceptancetests.tooling.{GraphServices, ServiceClient}
+import io.renku.graph.acceptancetests.tooling.{AcceptanceSpec, ApplicationServices, ServiceClient}
 import io.renku.graph.model.EventsGenerators.commitIds
 import io.renku.graph.model.testentities.generators.EntitiesGenerators._
 import io.renku.graph.model.{SchemaVersion, testentities}
@@ -34,14 +34,12 @@ import io.renku.http.client.AccessToken
 import io.renku.jsonld.syntax._
 import org.http4s.Status.Ok
 import org.scalactic.source.Position
-import org.scalatest.GivenWhenThen
 import org.scalatest.enablers.Retrying
-import org.scalatest.featurespec.AnyFeatureSpec
 import org.scalatest.time.{Minutes, Seconds, Span}
 
 import java.nio.file.{Files, Paths}
 
-class ReProvisioningSpec extends AnyFeatureSpec with GivenWhenThen with GraphServices with TSProvisioning with TSData {
+class ReProvisioningSpec extends AcceptanceSpec with ApplicationServices with TSProvisioning with TSData {
 
   Feature("ReProvisioning") {
